fix(data): remove casting to int in __getitem__
- casting caused errors when we had nans in concept labels

+++ concept_benchmark/data.py
@@ -1139,7 +1139,7 @@ class ConceptImageDatasetSample(ConceptDatasetSample):
             warnings.warn(f"{e}; cannot open image, returning path", RuntimeWarning)
             image = img_path
 
-        c = torch.from_numpy(np.array(c, dtype=np.int64))
+        c = torch.from_numpy(np.array(c))
         y = torch.from_numpy(np.array(y, dtype=np.int64))
 
         if self.concept_transform is not None:
